pyproject.toml
[project]
name = "my-fastapi-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.17.2",
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "fastapi[standard]>=0.122.0",
    "passlib>=1.7.4",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "redis>=7.1.0",
    "sqlmodel>=0.0.27",
]
